SMS_Server/v2/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
pyserial==3.5
gunicorn==20.1.0
python-dotenv==0.19.0
SQLAlchemy==1.4.23 